+++ server/src/static_file_filter.rs
@@ -14,7 +14,7 @@ use std::path::Path;
 use std::sync::Arc;
 use tokio::fs::File;
 use tokio::io;
-use tracing::debug;
+use tracing::{debug, trace};
 use warp::fs::{file_stream, optimal_buf_size, Cond, Conditionals};
 use warp::http::{Response, StatusCode};
 
@@ -202,15 +202,15 @@ pub async fn cache_or_file_reply(
                 //false,false => cache without content-encoding
                 //false, true => file
                 if !client_accept_gzip && compressed {
-                    debug!("{} hit disk", key);
+                    trace!("{} hit disk", key);
                     Ok(file_reply(&item, path.as_ref(), range, modified).await)
                 } else {
                     let mut resp = cache_reply(item.as_ref(), bytes, range, modified);
                     if client_accept_gzip && compressed {
-                        debug!("{} hit cache, compressed", key);
+                        trace!("{} hit cache, compressed", key);
                         resp.headers_mut().typed_insert(ContentEncoding::gzip());
                     } else {
-                        debug!("{} hit cache", key);
+                        trace!("{} hit cache", key);
                     }
                     Ok(resp)
                 }
@@ -255,7 +255,7 @@ pub async fn get_cache_file(
 ) -> Option<(String, Arc<CacheItem>)> {
     let _key = sanitize_path(tail)?;
     let key = _key[1..].to_owned();
-    debug!("get file: {host}, tail:{_key}, fixed: {key}");
+    //debug!("get file: {host}, tail:{_key}, fixed: {key}");
     if let Some(cache_item) = domain_storage.get_file(host, &key) { 
         Some((key, cache_item))
     } else {