+++ tests/src/main.rs
@@ -8,8 +8,9 @@ use tracing_subscriber::EnvFilter;
 async fn main() -> Result<()> {
     env::set_var(
         "SPA_CONFIG",
-        get_test_dir().join("server_config.conf").display().to_string(),
+        get_test_dir().join("server_config.toml").display().to_string(),
     );
+
     tracing_subscriber::fmt()
         .with_env_filter(
             EnvFilter::builder()